requirements.txt
streamlit==1.53.1
anthropic==0.76.0
gspread==6.1.4
oauth2client==4.1.3
PyYAML==6.0.2
python-dotenv==1.2.1
